# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/children": {
      "get": {
        "operationId": "ChildrenController_list",
        "summary": "List children",
        "description": "Paginated, filterable by status, groupId, and free-text search (q). Requires child:read; results are scoped to the caller's branch/group depending on their granted scope.",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Matches first, last, or middle name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasDebt",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasMedicalAlert",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "lastName:asc",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "children"
        ]
      },
      "post": {
        "operationId": "ChildrenController_create",
        "summary": "Create a child",
        "description": "Registers a new child record. Requires child:create.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChildDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new child is created with status `applicant` — it is NOT on any attendance screen until POST /children/:id/status moves it to `active` and it is assigned to a group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildSummaryDto"
                }
              }
            }
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}": {
      "get": {
        "operationId": "ChildrenController_get",
        "summary": "Get a child by ID",
        "description": "Returns full profile detail for a single child. Requires child:read.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "children"
        ]
      },
      "patch": {
        "operationId": "ChildrenController_update",
        "summary": "Update a child",
        "description": "Updates a child's profile fields. Requires child:update.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChildDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "children"
        ]
      },
      "delete": {
        "operationId": "ChildrenController_remove",
        "summary": "Delete a child",
        "description": "Permanently removes a child record. Requires child:delete, a sensitive permission — see the RBAC §5.5 safety rails (privilege-escalation and sensitive-permission gating).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/status": {
      "post": {
        "operationId": "ChildrenController_setStatus",
        "summary": "Change a child's status",
        "description": "Transitions a child between statuses (e.g. active, graduated, withdrawn) and records the transition in the status history. Setting status to withdrawn or graduated stamps the withdrawal date from dto.effectiveDate. Requires child:status.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/history": {
      "get": {
        "operationId": "ChildrenController_history",
        "summary": "Get a child's status history",
        "description": "Returns the full status-change history for a child. Requires child:read.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/documents": {
      "get": {
        "operationId": "ChildrenController_listDocuments",
        "summary": "List a child's documents",
        "description": "Returns metadata for every document on file for this child. Requires child:read.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      },
      "post": {
        "operationId": "ChildrenController_createDocument",
        "summary": "Add a document to a child's file",
        "description": "Attaches a document record (e.g. a scanned ID or medical form) to a child. Requires child:update. The document itself is uploaded separately via the files module.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChildDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/documents/{docId}": {
      "patch": {
        "operationId": "ChildrenController_updateDocument",
        "summary": "Update a child's document metadata",
        "description": "Edits an existing document record for this child. Requires child:update.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChildDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/documents/{docId}/verify": {
      "post": {
        "operationId": "ChildrenController_verifyDocument",
        "summary": "Verify a child's document",
        "description": "Marks a document as staff-verified (e.g. an ID checked against the original). Requires child:update.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/medical": {
      "get": {
        "operationId": "ChildrenController_getMedical",
        "summary": "Get a child's medical record",
        "description": "Requires medical:read — a separate permission from general child:read.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      },
      "put": {
        "operationId": "ChildrenController_updateMedical",
        "summary": "Replace a child's medical record",
        "description": "Full replace (PUT, not PATCH) of medical data — allergies, conditions, medications, etc. Requires medical:write.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMedicalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/medical/alerts": {
      "get": {
        "operationId": "ChildrenController_medicalAlerts",
        "summary": "Get a child's medical alerts",
        "description": "Returns the short list of critical medical flags (e.g. severe allergies) meant for staff to see at a glance, separate from the full medical record. Requires medical:alerts.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/consents": {
      "get": {
        "operationId": "ChildrenController_listConsents",
        "summary": "List a child's consents",
        "description": "Returns recorded parental consents (e.g. photo/media release, field trips) for this child. Requires consent:read.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      },
      "post": {
        "operationId": "ChildrenController_recordConsent",
        "summary": "Record a consent",
        "description": "Records a new parental consent for this child. Requires consent:manage.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordConsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
    "/api/v1/children/{id}/consents/{consentId}/revoke": {
      "post": {
        "operationId": "ChildrenController_revokeConsent",
        "summary": "Revoke a consent",
        "description": "Revokes a previously recorded consent. Requires consent:manage.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "consentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "children"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kindergarten API",
    "description": "Kindergarten Management System — backend API (Stage 1)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChildSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "example": "Alisher"
          },
          "lastName": {
            "type": "string",
            "example": "Karimov"
          },
          "middleName": {
            "type": "object",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "2021-04-17"
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "enum": [
              "male",
              "female"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "applicant",
              "active",
              "temporarily_absent",
              "suspended",
              "graduated",
              "withdrawn",
              "archived"
            ],
            "description": "A child created via POST /children starts as `applicant` — it must be moved to `active` before it appears on attendance screens."
          },
          "photoFileId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "groupId": {
            "type": "object",
            "nullable": true,
            "format": "uuid",
            "description": "Current group (group_assignment with effectiveTo = null); null if unassigned."
          },
          "groupName": {
            "type": "object",
            "nullable": true,
            "example": "Quyoshcha"
          }
        },
        "required": [
          "id",
          "branchId",
          "firstName",
          "lastName",
          "birthDate",
          "status"
        ]
      },
...
      "ChildListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateChildDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ]
          },
          "address": {
            "type": "string"
          },
          "enrollmentDate": {
            "type": "string"
          },
          "contractNumber": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "branchId",
          "firstName",
          "lastName",
          "birthDate"
        ]
      },
...
      "ChildDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "example": "Alisher"
          },
          "lastName": {
            "type": "string",
            "example": "Karimov"
          },
          "middleName": {
            "type": "object",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "example": "2021-04-17"
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "enum": [
              "male",
              "female"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "applicant",
              "active",
              "temporarily_absent",
              "suspended",
              "graduated",
              "withdrawn",
              "archived"
            ],
            "description": "A child created via POST /children starts as `applicant` — it must be moved to `active` before it appears on attendance screens."
          },
          "photoFileId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "groupId": {
            "type": "object",
            "nullable": true,
            "format": "uuid",
            "description": "Current group (group_assignment with effectiveTo = null); null if unassigned."
          },
          "groupName": {
            "type": "object",
            "nullable": true,
            "example": "Quyoshcha"
          },
          "address": {
            "type": "object",
            "nullable": true
          },
          "enrollmentDate": {
            "type": "object",
            "nullable": true,
            "format": "date"
          },
          "withdrawalDate": {
            "type": "object",
            "nullable": true,
            "format": "date"
          },
          "contractNumber": {
            "type": "object",
            "nullable": true
          },
          "registrationNumber": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "childGuardian": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildGuardianLinkDto"
            }
          },
          "allergy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildAlertDto"
            }
          }
        },
        "required": [
          "id",
          "branchId",
          "firstName",
          "lastName",
          "birthDate",
          "status",
          "childGuardian",
          "allergy"
        ]
      },
      "UpdateChildDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ]
          },
          "address": {
            "type": "string"
          },
          "enrollmentDate": {
            "type": "string"
          },
          "contractNumber": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "photoFileId": {
            "type": "string"
          }
        }
      },
      "ChildStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "applicant",
              "active",
              "temporarily_absent",
              "suspended",
              "graduated",
              "withdrawn",
              "archived"
            ]
          },
          "effectiveDate": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "effectiveDate"
        ]
      },
      "CreateChildDocumentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "birth_certificate",
              "parent_passport",
              "medical_certificate",
              "contract",
              "consent",
              "other"
            ]
          },
          "number": {
            "type": "string"
          },
          "issueDate": {
            "type": "string"
          },
          "expiryDate": {
            "type": "string"
          },
          "fileId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "UpdateChildDocumentDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string"
          },
          "issueDate": {
            "type": "string"
          },
          "expiryDate": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "UpdateMedicalDto": {
        "type": "object",
        "properties": {
          "bloodType": {
            "type": "string"
          },
          "chronicConditions": {
            "type": "string"
          },
          "emergencyInstructions": {
            "type": "string"
          },
          "doctorName": {
            "type": "string"
          },
          "doctorPhone": {
            "type": "string"
          },
          "clinic": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "allergies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllergyDto"
            }
          },
          "medications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicationDto"
            }
          }
        }
      },
...
    }
  }
}